[REF] base_vat: insert guidelines
* Add guidelines for tests and references to manage patches to
  stdnum
* Remove outdated patches to stdnum (ie, mx, nl, no, pe, ru), already
  present in version 1.13
* Make the methods a mixin so that it is less awkward on fiscal
  positions
* Raise better errors with hints
