+++ addons/base_vat/models/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import base_vat_mixin
+from . import countries
 from . import account_fiscal_position
 from . import res_config_settings
 from . import res_company
