+++ addons/base_vat/tests/__init__.py
@@ -1,2 +1,2 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import test_validate_ruc
+from . import test_validate_vat
