+++ addons/base_vat/views/res_partner_views.xml
@@ -5,8 +5,14 @@
             <field name="model">res.partner</field>
             <field name="inherit_id" ref="base.view_partner_form"/>
             <field name="arch" type="xml">
-                <xpath expr="//field[@name='vat']" position="attributes">
-                    <attribute name="string">VAT</attribute>
+                <xpath expr="//field[@name='vat']" position="before">
+                    <div class="alert alert-warning p-1" role="alert" attrs="{'invisible': [('vat_error', '=', False)]}" colspan="2">
+                        <label for="vat_error" string=""/>
+                        <field name="vat_error"/>
+                    </div>
+                </xpath>
+                <xpath expr="//label[@for='vat']" position="replace">
+                    <field name="vat_label"/>
                 </xpath>
                 <xpath expr="//span[hasclass('o_vat_label')]" position="replace">
                     <span class="o_vat_label">VAT</span>
