+++ addons/l10n_ae/data/l10n_ae_data.xml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <!-- set VAT label to show on invoice report -->
-    <record id="base.ae" model="res.country">
-        <field name="vat_label">VAT</field>
-    </record>
     <record id="base.AED" model="res.currency">
         <field name="symbol">AED</field>
     </record>
