--- addons/l10n_sa/data/account_data.xml
@@ -1,8 +1,0 @@
-<odoo>
-    <data noupdate="1">
-        <!-- set VAT label to show on invoice report -->
-        <record id="base.sa" model="res.country">
-            <field name="vat_label">VAT Number</field>
-        </record>
-    </data>
-</odoo>
