+++ odoo/addons/base/data/res_country_data.xml
@@ -13,6 +13,7 @@
             <field name="code">ae</field>
             <field name="currency_id" ref="AED" />
             <field eval="971" name="phone_code" />
+            <field name="vat_label">VAT</field>
         </record>
         <record id="af" model="res.country">
             <field name="name">Afghanistan</field>
@@ -1227,6 +1228,7 @@
             <field name="code">sa</field>
             <field name="currency_id" ref="SAR" />
             <field eval="966" name="phone_code" />
+            <field name="vat_label">VAT Number</field>
         </record>
         <record id="sb" model="res.country">
             <field name="name">Solomon Islands</field>
